refactor add command and functionality to use new cmd/subcmd architecture
